Flatten the wired balance box styles
